Add guardrails to queue list query and index coverage

+++ crates/sr-common/src/db/match_results.rs
@@ -17,8 +17,8 @@ pub enum MatchResultStorageError {
 
 #[derive(Debug, Clone, Default)]
 pub struct MatchResultInsert {
-    pub talent_id: i32,
-    pub project_id: i32,
+    pub talent_id: i64,
+    pub project_id: i64,
     pub is_knockout: bool,
     pub ko_reasons: Option<Value>,
     pub needs_manual_review: bool,
